Initial commit!
See README for a very broad overview.

Right now, I've implemented the phone-based layout, with tabs for each
part of a section. This means that the three fragments can be
constructed, managed, and given data; this is nicely independent from
the code that manages the layout - the layout creates the fragments and
places them in the heirarchy, but doesn't mess around with their
internals. The fragments don't do anything to the layout code - they
simply call back to the activity.

Lots of stuff  left to do:
-finish the link-sequence browser (it doesn't show anything yet).
-cause activity launching for link selections.
-implement the multi-pane tablet layouts. Yes, layouts, plural.
-load and store of the document model.
-settings, to allow selecting documents. Maybe multiple docs, w/ some
present in side menu?
-add the editing modes, to allow adding, reordering, renaming, etc.
-create activity (w/ dialogs) for adding links, e.g. by external req.