Overhauls, one major, one minor.
Realized a major problem with layout code: it couldn't got back
properly. I ended up implementing a browse stack in LayoutMgr. This
allows the mgr to store references to the entire chain of fragments that
lead to a specific location. This also allows subclasses even more
flexibility in deciding what to do with a section - for instance, the
new TabLayout2 design will only display tabs for a section's non-empty
collections - it won't display a tab for subsections if there are no
subsections. Since the mgrs can now determine whether or not they're at
the root, it's easier for them to set up the stuff differently depending
on whether or not they're at root. Also, all the code and interfaces
become much cleaner.

Other change involves the Async stuff, which now can load either from
the sdcard or from raw resources. Raw resource loading will let me
bundle in a default document for new users. Split stuff into separate
traits to make it cleaner, and also easier to implement new kinds of
loaders later on (I'm thinking maybe loading from web, for instance.
We'll see.)