Guess what! Another overhaul!
Second one in as many days. Decided to eliminate fragments entirely (you
can see that the browser/ package is gone). The activity now manages a
path stack itself, and the layouts now involve listviews directly. The
trick is that the managment involves creating new list adapters and
slotting them in.

Now, this is very elegant and all, but why break a good thing, you might
ask. Well, it's because this way I can use the mergeadapter. I've
already applied the merge adapter to the new link sequence adapter,
where it works quite nicely. The big plan is to use it to join the links
and link sequences lists together in the tablet layouts, so they can
fill the screen nicely.

Also, I've currently taken out the Subdisp layout; I might repair it for
the new LayoutMgr spec eventually. It's not as important now that I've
got the tabbed version working.