Fixes #2 by implementing instance state saving.
Loading system has been modified so that it gets passed callbacks for
both success and failure (I am thinking ahead to #1). See Async.scala.
Types have been defined for the callbacks to make signatures easier.

Created a type (DocLocation) and cases classes for references to files,
depending on where they are. I've also created a load tool that knows
which doc loader to use based on the passed in DocLocation.

I've implemented onSaveInstanceState; it puts the current doc loc info,
and the path, in the bundle. onCreate now first looks for info in the
the saved instance state, then calls method for #3 (user files list),
then defaults to some default file (c.f. #5).

Also added methods to LayoutMgr for saving and restoring state. Bundles
come in through the callback closure in LSA's loadDoc. This is
especially useful rotating the screen on small devices, where TabLayout2
is the mgr for both orientations - TabLayout2 uses the mechanism to make
sure that the same tab is selected on the return as was before.