Both tablet layouts done. Fixes #9.
Coding was fairly straightforward, given the fix in the previous
commit. Basically, the links and link sequences are in a single list,
using a merge adapter. The listener directs the proper, translated info
to the passed in callbacks. Very simple.

The tricky part is the layout stuff. It's kind of a mess, making me
basically terrified of #8. Also, I think the scrollbars are off. Good
news is, I've added drop shadows!

THere are some fixes that'll need to be done later, but the point is
clear.